requirements.txt
pydantic==2.6.4
uvicorn==0.29.0
numpy
pandas
requests==2.31.0
openai==1.16.1
python-dotenv==1.0.1
fastapi==0.110.1
tiktoken==0.5.2 
nest-asyncio==1.6.0 
aiohttp==3.9.5
transformers==4.40.0
huggingface-hub==0.22.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
fastparquet
torch==1.13.1+cpu
torchvision==0.14.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html